Adding transient object labels on mouse hover.  To detect mouse hovering, KStarsData::updateTime() calls SkyMap::checkHoverPoint(), which checks to see if the mouse has been stationary for >0.5 sec.
If a hover event is trigered, it searches for the nearest
object to the mouse cursor, sets the TransientObject pointer
to that object,  and adds a label with color UserLabelColor.

checkHoverPoint() also determines when the mouse is moved,
thus ending the hover event.  In this case,
fadeTransientLabel() is called, which slowly fades the label
into SkyColor and then sets TransientObject to NULL.

In implementing this feature, I added
SkyMap::objectNearest( SkyPoint* ), which returns a pointer to
the SkyObject nearest the agument.  This code used to be inside
SkyMap::mousePressEvent().

Please review the new functions (especially
SkyMap::checkHoverPoint(), and SkyMap::slotTransientTimeout() ).

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=285588