Added to the KStarsPopupMenu (triggered by clicking the right mouse button on the sky map) a label which shows the distance from the Earth to the sky object.  This information is in the Details entry and is coded in class DetailDialog.  If the object is a star the distance is shown in parsecs, if it is a planet or it is the sun it is shown in Astronomical Units. There is no distance information for asteroids yet.
  There are a couple of pending issues. The distance shown in the label
is obtained from the parallax of the object in the Hipparcos catalog. It
is not clear which is the smallest believable parallax. ESA says that
the Hipparcos catalog has a median astrometric precission of 7 mas
for magnitude < 9 and 25 mas for magnitudes > 10-11. However there
are many stars in the catalog with parallax < 1 mas and an error in
parallax lower than 1 mas. The manual for the catalog says that parallax
is realistic if parallax/delta parallax > 3. However we cannot check
this condition because we are using a customized Hipparcos catalog without
parallax errors.

  I have placed a limit of 0.5 mas as the smallest believable parallax
(corresponds to 2000 pc), but this is somehow arbitrary. If distance is
greater than that value one gets a message saying that distance is > 2000 pc.
Another way to estimate distance would be to get it from the relative and
absolute magnitudes. We will leave that for the future....-)

  The second issue is concerning localization of values for magnitudes and
distances. We are using QLabels to show the values. If we used KLineEdits
we could use the localization. How can the localization be used in a QLabel
as it is in KLineEdit?.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=288245