Implemented Manual stepping of the clock (wish #65476).
Press the "." / ">" key to advance one step forward; press the "," / "<"
key to advance one step backward.  The stepsize is set by the TimeStep
widget in the toolbar.

If the clock is running, it will be automatically stopped when one of the
manual-step keys is pressed.

On my keyboard, "," and "<" appear together on the same key, and "." and
">" are on another key; is that generally true for all keyboards?

At this point, there is no GUI way to do the manual-stepping.  It would
make sense to add these buttons to the TimeStep widget, but it is already
pretty complicated.  What do you think?

CCMAIL: 65476-done@bugs.kde.org
CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=289610