Modified the calculator's horizontal coordinates module (this was added to the wish list by Jasem ..:-) )
Now it is possible to select the input coordinates and perform transformations
of coordinates from catalog coordinates (or also from apparent
coordinates) to horizontal coordinates or the other way round, from horizontal
coordinates to apparent coordinates. If we go from horizontal coords to
equatorial apparent coordinates the epoch is considered to be the one
defined by the date and time in the upper group box. There is implementation
to choose another epoch. That is, I do not allow to go from horizontal
coordinates to a reference epoch like J2000 or 1950. If you need to do that,
you may choose the date and time in the upper group box for January 1st 2000
and you will get coordinates which are very similar the catalog ones for J2000.

I have also removed the epochToJd function since this function is already
defined in KSUtils so that it can be used from different calculator modules.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=290010