Introducing the kstars_gl branch, in which OpenGL support will be developed.  To checkout this branch, use "cvs co -r kstars_gl kdeedu/kstars".
I branched this from HEAD; the only differences at this point are in
SkyPoint and SkyMap.  If you have OpenGL working, you can test it out.
However, it is currently very rudimentary:  it displays 5000 bright stars,
the coordinate grid, and the constellation lines, using Equatorial
Coordinates.  You can slew the map with the arrow keys and zoom in/out
with the +/- keys.  Note that FocusPoint is not currently defined, so the
statusbar coordinates, and identification of objects with mouse clicks
will not work properly.

SkyPoint:  Now has double X,Y,Z members and accessor methods.  Wherever
the SkyPoint's position is changed, I have added "setXYZ()", which
re-synchronizes the XYZ coordinates to the new RA/Dec coordinates.

The XYZ positions are not volumetric; that is, they do not take distance
into account.  They are simply the XYZ coordinates on the suface of the
celestial sphere.  I think doing volumetric XYZ is going to be
complicated, because the size/brightness of objects will decrease with
distance, and our magnitudes/sizes already take distance into account.
Plus, for many objects, we do not know the distance.  These are solvable
problems, of course.

SkyMap: I have replaced the paintEvent() function with its QGLWidget
equivalent, paintGL().  In that function, we now call GL-equivalents of
each of our draw functions.  Many of these have not yet been implemented,
and remain commented out.

In addition, I have reimplemented the arrow-key and +/- key cases in
SkyMap::keyPressEvent().

I welcome all interested parties to join me in developing this branch,
either by commiting code or making suggestions on the list!

CCMAIL: kstars-devel@kde.org

svn path=/branches/kstars_gl/kdeedu/kstars/; revision=290278