DetailDialog improvements:
Added Angular Size to "General Info" box, and added
Hour Angle and Airmass to "Coordinates" box.

Fixed layout of "General info" box (the left column
stretches now instead of the right)

When a data field (magnitude, distance, or angular
size) is unknown, it is now consistently rendered
as "--" (we used to use "?" and "unknown" as well).

Now using the dms::to[D|H]MSString() functions
instead of building the coordinate strings manually.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=291514