Converting KStars to the KConfigXT framework, which will make management of our many config options much easier.
With this framework, all of our options are described in the
XML file kstars.kcfg.  The build process parses this file
and generates an Options class with accessors and mutators
for each option.  Using Qt Designer, the widgets controlling
each option are named "kcfg_OptionName", and then these are
automatically connected to the corresponding option with no
actual coding required.

In practice, where you used to see "options()->optionName",
you will now see "Options::optionName()" for access and
"Options::setOptionName(newValue)" for mutation.

To add a new option, just describe it in the kstars.kcfg
file, and make sure to name its widget controller (if any)
in Qt Designer with the "kcfg_OptionName" scheme.


Notes:

I added a "Colors" group to the kstarsrc file.

There is no longer a viewopsdialog class.  Instead, we just
create a KConfigDialog object and add the Ops* widgets as
pages.  Almost all of the code that was in ViewOpsDialog is
not necessary under KConfigXT, but some of it we still
need for controlling complex widgets that KConfigXT can't
deal with:
- the listbox of deep-sky and custom object catalogs
  (now in OpsCatalog.cpp)
- the listboxes of adjustable colors and predefined color
  schemes (now in OpsColors.cpp).

Because it is now a standard KConfigDialog, the View Options
window no longer does "live" updating of the skymap as
options are changed.  It does have an "Apply" button,
which achieves the same result, just not automatically.

It seems to be working pretty well; please test the behavior
of options if you can.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=292754