Make the "Apply" button in the View Options dialog actually work as it should (update the skymap with new settings applied when pressed).
Also, two widgets (one in opsguides, one in opsadvanced) were named
incorrectly due to typos.  One of these was actually causing a crash on
my work machine, but that may be due to the fact that it has KDE 3.2-rc1,
since it wasn't crashing on my home machine.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=293013