The KStars Setup Wizard now has some actual functionality. You can set the Geographic Location, and launch the telescope setup wizard. I also added a "Download Extra Data" page, but it has no functionality yet.
TODO: get the extra data page working; add a page to set some common preferences; have Wizard run automatically on first run of KStars.

I also fixed a bug with the "slewing timescale"; the timescale above which the simulation engages the clock's "manual" mode, which ticks the clock as often as possible, rather than at regular intervals. Because kconfigXT doesn't know how to interface with the custom timestepbox widget, it was unabe to control this parameter in the OpsAdvanced page of the Configuration window. This widget is now handled manually.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=296190