The background of the on-screen InfoBoxes can now be drawn semi-transparently or opaquely, using the (previously unused) BoxBGColor color.  The default background is semi-transparent.  In the Colors tab of the Options window, there is a combobox where you can select from "no fill", "transparent", or "opaque".  This is not real alpha-transparency; I'm just using a dense dither pattern (Qt::Dense4Pattern).
I also adjusted the BoxBGColor of the default color scheme.

Right now, the BG mode value is stored in the Options object, but it is
not part of the ColorScheme class.  I couldn't decide whether it should
be there or not.

I also fixed a crash condition, when the Horizon label was drawn a few
degrees west of the North horizon point.  Also, the Horizon label is no
longer drawn when the Opaque ground is active.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=297065