Fixing faint limits of deep-sky objects.  The spinboxes in the options window were limited to 10.0 mag, but many DSOs are fainter than this. The spinboxes now go up to 16.0.
However, because the magnitudes of the NGC/IC catalog are so incomplete
(and the ones that are there I have found to be unreliable), I am
disabling an adjustable faint limit for DSOs.  I will re-enable it once
the magnitudes of the ngc/ic objects can be completed and corrected.
Any volunteers?  ;)

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=302944