These changes are necessary to fix a bug in the Solar system viewer which shows in wrong position the planets, mainly the inner ones, since they are depicted according to their ecliptic geocentric coordinates and not according to their heliocentric ones.
I think that the KSplanetBase class should only hold the heliocentric ecliptic coordinates
of the planets, while the geocentris ecliptic coordinates, without the distance either to the sun
to the to earth, should be stored in SkyPoint. However I have not yet implemented this change
because it requires too many modifications and before we need to agree in KStars-devel.
Therefore I have implemented the simplest solution:

- Added a private member called HelEcPos which stores the Helicentric ecliptic coordinates
  of the planets.
- Added six methods to retrieve and set the heliocentric ecliptic coordinates of the planets:
  helEcLong() and helEcLat() return the Heliocentric ecliptic longitude and latitude
  respectively referred to the mean equinox of the date. setHelEcLong() and setHelEcLat()
  set the Heliocentric ecliptic longitude and latitude respectively and are overloaded to
  accept a double or a dms.

The heliocentric ecliptic coordinates are set in KSPlanet class when calling the
findGeocentricPosition function.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=305481