Added a new calculator module which computes the Start Date and Time of the equinoxes and solstices, and the duration of the selected season for a given year.
Present status:
- The interactive mode works fine. There is an error of 2 minutes for the spring
  equinoxes (I am not very happy with this) and errors of at most 10 seconds for the
  autumn equinox and the winter and summer solstices. I still do not know why and
  will investigate it further. However it should be noted that an error of 1 arcsec in
  the position of the sun causes an error of 24 seconds in time.
- The file mode is not yet operative although I have already designed the GUI dialog and the
  module shows it. I will complete it as soon as possible.

In order to make these computations I have added a static function to KSUtils which performs
Langrange interpolations with at most 10 points. I have added some functions to KSSun which
compute the Julian Day for the equinoxes and solstices. This is done by finding the julian day
at which the apparent ecliptic (geocentric) longitude of the sun is 0, 90 180 and 270 degrees.
To compute this value I interpolate using 5 values around the expected dates and the Lagrange
algorithm defined in KSUtils. Here there is a small issue. I modify permanently the position
of the sun after making the computations. I need some help from the kstars-devel list to
find a cleaner solution for that. On the other hand this will not affect the normal operation
of KStars, because it only affects the KSSun object which I create in that calculator module
and I destroy it after getting out of this module.

I have still to investigate if I have to make a very small correction to convert from the epoch
obtained by the VSOP computations, it is referred to the dynamical ecliptic and equinox, and
the standard FK5. I guess this correction would ammount approx. 3 or 4 seconds in time.

CCMAIL:kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=305964