Two SimClock fixes:
manualTick(): SimClock stores the timescale as a double,
representing the size of the timestep in seconds.
manualTick() used to advance the clock by calling
setUTC( UTC().addSecs( int( Scale ) ) );

This did not work if the Scale was larger than the range
of integers.  We now divide Scale by 86400. to get the
integer number of days, and then call addDays(ndays) and
addSecs( int(Scale - ndays*86400) ) to advance the clock
correctly.

setUTC(): Do not set the time if the requested ExtDateTime
is invalid.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=310917