Two fixes:
1.  Transient labels now work even when the clock is stopped.
I am now using an independent QTimer to determine when the
mouse remains unmoved for 500 ms, which causes it to assign
the TransientObject pointer to the object nearest the mouse.
(I used to assign the TransientObject pointer after 5 cycles
of updateTime(), which is why it did not work with the clock
stopped).  Also, at Pablo's suggestion, transient labels are
now attached to unnamed stars.

2.  Added functions to SkyObject:  translatedName2() and
translatedLongName().  We already had translatedName(), so I
added these for consistency and to make sure all object names
get localized.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=312232