Now KStars can save FITS in 8, 16, 32, and IEEE -32, -64 modes. It used to save only in 8bits.
Also, I'm in the process of converting all FITSViewer tools to handle a channel depth higher than 8bits. For example, the contrast/brightness tool can be used now to alter the contrast/brightness of any arbitrary channel, making it possible to perserve values in FITS files with depths of 16bits or higher.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=315976