Added a new module to the astronomical calculator which computes the angular separation between two points in the sky whose coordinates are given in equatorial coordinates. Both modes, interactive and batch mode are fully implemented and operative.
Right now it is not possible two enter the coordinates of the points in a different coordinate
system (like the horizontal one), because the module assumes that the "X" or "alpha" coordinate
for both points is given in hours:minutes:seconds and not degrees:arcmin:arcsecs.

I have housed this module in the "Coordinate Converters" section, although I am not very satisifed
with this choice. It might have been better to create a new section for miscellanea computations.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=316553