Adding the 4 new DCOP functions (exportImage, printImage, loadColorScheme, setColor) to the ScriptBuilder tool.
Adding width and height parameters to the exportImage DCOP function.  If the width and height parameters are set to zero, then the exported image will use the dimensions of the sky map widget.  Note that the skymap image is not scaled to match the requested image size.  If the requested image is smaller than the skymap image, it will be cropped.  If the requested image is larger than the skymap, then the exported image will be padded with white pixels.

Fixed bug:  when switching color schemes, the starColorMode and starColorIntensity were not being updated.  Fixed in KStars::loadColorScheme().

(PS: I posted another commit earlier, but have not seen it appear on the mailing list.  In case I got the addres wrong: that commit was in libkdeedu/extdate, and it fixes the crash condition in the ScriptBuilder tool)

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=316669