The Details Dialog now displays the angular size of major solar system bodies (i.e., not comets and asteroids).  The planets' angular sizes are shown in arcseconds, while the Sun and moon use arcminutes.  Also, the distance to comets and asteroids is now shown, and the distance to the Moon is now displayed in km rather than showing "0.0 AU" :).
To include the angular sizes, I added code to compute the real angular size of
each planet, based on its physical diameter, and its distance from Earth.
Previously, I had just hard-coded an "average" angular size that was really
little more than a rough guess.  The physical diameter is now an argument in
the KSPlanetBase ctor and those of its children.  The diameters are defined in
PlanetCatalog::initialize(), except for the Moon (which is defined in
KSMoon::KSMoon() ).

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=317998