Fixes to layout of calculator modules.  All dmsBoxes now have a minimum width of 100 pix.  There were some group boxes that did not have a layout.  I rearranged widgets in the new planets module a bit.
One remaining issue:  In the Sidereal Time module's batch mode, if you reduce
the width of the window, the widgets in the top group box overrun the box edge.
I don't understand why this is occurring; that box has a layout, and in fact I
tried several different layout strategies, but they all had this problem.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=318500