Removing compiler warnings.  Other than INDI, we now have only one compiler warning: "skypoint.cpp:229: warning: unused parameter `bool includePlanets'"
includePlanets is a parameter for the function SkyPoint::updateCoords().
It is unused in this function, but is used in the overloaded
KSPlanetBase::updateCoords().  I could add some trivial use of
includePlanets to SkyPoint::updateCoords(), but since this function
is called quite frequently, I would rather have the warning message.
Does someone know another way to silence the warning?

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=320453