Fixing rise/set times.  Joseph North pointed out that the Moon's set time was sometimes given for the following day, rather than the current day.
This is actually a complicated issue.  To compute the "set time", you
first have to specify *which* set event you are talking about.  One can
think of several sensible ways to define it:  the next chronological set
event; the set event which occurs on the current calendar date; the
nearest set event going either forward or backward in time.  It isn't
clear that any of these is obviously the best idea, yet they all behave
differently.

We used to simply display the previous rise event, and the next set
event.  Now, we first assess if the object is currently below the
horizon.  If so, we modify the time by +/- 12 hours to bring the object
above the horizon, and *then* calculate the previous rise event and next
set event.

This gives us the behavior that the rise/set times only change at the
instant an object crosses the lower meridian (when it reaches its lowest
point and changes from "setting" to "rising").

Also, I improved the behavior in the WUT tool.  It was also giving
rise/set times for dates which were sometimes off by one.  We now use
special reference times (midnight, 6:00 in the evening, and Noon the
following day) to ensure that the various rise/set times are computed
correctly.

I noticed that the transit times are completely bogus.  I'm looking into
it.

CCMAIL: kstars-devel@kde.org
CCMAIL: joseph.north@mail.mcgill.ca

svn path=/trunk/kdeedu/kstars/; revision=331083