Attempting to fix some dialog layouts.  The Manual Focus tool's widgets did not stretch when the window was resized.  This is fixed. The What's Up Tonight tool's widgets stretch, but the window does not have a minimum size: it can be sized to much smaller than the widgets it contains.
I am unable to fix this; AFAICS, there is no difference between
the layout code in WUTDialog's ctor:

	QFrame *page = plainPage();
	setMainWidget(page);
	QVBoxLayout *vlay = new QVBoxLayout( page, 0, spacingHint() );
	WUT = new WUTDialogUI( page );
	vlay->addWidget( WUT );

and that in, say, FocusDialog:

	QFrame *page = plainPage();
	setMainWidget(page);
	vlay = new QVBoxLayout( page, 0, spacingHint() );
	acd = new AddCatDialogUI(page);
	vlay->addWidget( acd );

Yet the two dialogs behave very differently: the FocusDialog window is
constrained to be as large as the widgets within it.  The difference
must be somewhere else, but I can't find it.  Interestingly, if you load
kstars/kstars/tools/wutdialogui.ui in designer, then it *is* properly
constrained to be as large as its widgets.

Any ideas on this?

CCMAIL: kstars-devel@kde.org
CCMAIL: kde-edu@kde.org

svn path=/trunk/kdeedu/kstars/; revision=331372