More fixes for Rise/Set times: The popup menu set time was still given for the following day, rather than the current day.
Also, there was sometimes a 1-minute discrepancy between the times in
the popup menu and the times in the WUT tool.  This was because the
popup menu times were rounded to the nearest minute, while the WUT was
not rounded.  Both are rounded now, and the handling of it is
simplified.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=333428