Fixing scaling and rotation of symbols and images when printing the sky image.
There were missing "scale" paameters in some draw functions, which was okay
for screen draws because the default value of 1.0 is correct in that case.
Not so for printed images.

Also added a separate fix to rescale the star pixmaps properly on the printed
image.

Fixes to SkyMap::findPA() to correctly determine an object's position angle

Made SkyMap::pa() virtual so that the overridden versions in KSPlanetBase and
DeepSkyObject will be picked up correctly.

With these fixes, the Print function works again (these problems were all
regressions)

Oh, and in case you're wondering how slow it is to bugfix through an SSH
tunnel that spans continents, the answer is: very, very slow.  :)
I was able to work on this because the wind got too strong last night, so we
had to close the telescope.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=334906