Experimental usability improvement for dmsBox:
When the dmsBox is empty, it displays in grey text
"dd mm ss.s" or "hh mm ss.s" (depending on whether it
expects degrees or hours).  These letters disappear
when the dmsBox gets input focus.  They reappear if
the dmsBox loses input focus and the box is empty.

The intention is to provide the user with some hint
that this isn't just a normal edit box, that angle
values can be accepted by typing values for degrees,
arcminutes and arcseconds (or hours, minutes, seconds).
It also provides visual feedback on whether degrees
or hours are expected.

Drawbacks: it may be confusing to some users who
are not familiar with expressing angles in this
way.  It also does not make clear that other formats
are also acceptable.

Please let me know your opinions on this feature.
You will need to run "make -f Makefile.cvs" and
configure, because dmsBox now needs a MOC file.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=351307