Fixing BR #89290 (Machholz 2 text displayed twice slightly offset)
This isn't really a bug.  Comet Machholz 2 was discovered in 1994, and
at that time it consisted of 4 separate nuclei on a common orbit.  It
must have been broken up, either from internal forces of its outgassing
or due to tidal forces from the Sun or Jupiter.

Since 1994, two of the nuclei have apparently gone dark, but the other
two are still observable.  Since they are on a common orbit, they appear
very close in the sky, which is why you see the overlapping text of
their common name.  See here for more info:
http://cometography.com/pcomets/141p.html

To avoid confusion, I have changed their names to include the "A" and
"D" designations for the separate nuclei.

(Note that comet Schwassmann-Wachmann 3 has four separate components in
KStars, but these already had different designators on each nucleus)

CCMAIL: 89290-done@bugs.kde.org
CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=351970