Improvements to the Solar System viewer:
1) Orbital paths are now the true ellipses instead of circles.
2) Can move around the plot with the arrow keys

The orbital paths are constructed by recomputing the actual position of
each planet for 100 places along each orbit.  This takes a few moments,
so I use the "delayed initialization" trick to show the (blank) window
first, and then do the calculations.  We may want to just store the
orbit geometry for each planet in a data file to make this go faster.

Added code to compute the heliocentric ecliptic coords of Pluto.

TODO:

+ Add comets and asteroids, moon, and Jupiter's moons
+ Add an independent simulation clock with timestep widget
+ Add "center and track" on planetary bodies using double-click and
  0-9 keys
+ Add ability to rotate point of view angles

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=353583