Fulfilling wish #82946 (include constellation in object details).
The constellation which contains the object is now displayed in the
details dialog, and also in the top section of the popup menu.  Please
let me know if you think the latter is overkill.  I kind of like it, but
I know some people prefer smaller popup menus :)

Added function SkyPoint::constellation(), which returns the name of the
constellation which contains the point.  It does this by iterating over
all of the constellation-boundary line segments to find the pair of
segments which bracket the point to the North and South.

FEATURE: 82946
CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=357219