Fixing bug #96260 (Angular ruler line doesn't update when moving with arrow keys).
The Angular ruler geometry is now changed in a new function,
SkyMap::updateAngleRuler().  This is called in drawOverlays() if
isAngleMode() is true.  The result:  the angular ruler line is updated
whenever the sky is changed (by mouse, arrow keys, the passing of time,
etc.)

(BTW, to the bug reporter: your report implies that the line appears at
startup, which should not happen.  That's an old bug that has been
fixed)

CCMAIL: 96260-done@bugs.kde.org
CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=375456