(Didn't we do this once before? I vaguely remember dealing with timezone problems before).
The integer variable timezone is a Linuxism. Introduce a static
inline function that returns the right thing also on non-Linux
platforms.

The use of variable daylight is a little confusing to me: on
non-Linux systems it now reflects whether DST is in use _now_,
while the Linux documentation says:

       GMT) and daylight (0 if this time zone does not have any daylight  sav-
       ings  time  rules, nonzero if there is a time during the year when day-
       light savings time applies).

so it is probably nonzero in nearly every timezone nearly all the time
under Linux.

svn path=/trunk/kdeedu/kstars/; revision=375680