Fixes to User Log tab of Details dialog (BR #96670):
(1) Removed Save button.  Instead, I subclassed KTextEdit to emit a
signal when it loses input focus, and this signal is connected to the
saveLogData() slot.

(2) The user log text was being saved to disk, but was not being written
to the SkyObject's userLog member variable, so newly entered log text
was not visible unless KStars was restarted.

BUG: 96670
CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=377359