Command line changes.
(1) The "--date" argument is now applied when starting in normal GUI
mode (it used to only be read when using image-dump mode).

(2) There is now a "--paused" argument which will make KStars start with
the clock paused.  This fixes BR #78068, and should help people on
slower machines.

So, to start KStars at 12:30 on October 3rd, 2002 with the clock paused,
start it like:

% kstars --paused --date "3 Oct 2002 12:30:00"

BUG: 78068
CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=377366