Added a private function findMagnitude(KSNumbers *num) which computes the visual apparent magnitude (m_v) of the planets in class KSPlanetBase. The algorithm for computing it from the phase angle has been provided by Pere Planesas (OAN) for KStars. I am using a fixed distance between the Sun and Earth of 1 AU, instead of computing it to obtain the phase angle. The phase angle is obtained following Meeus book "Astronomical Algorithms".
Function findMagnitude() is called within function findPosition() and therefore
is executed for all solar system bodies. That means that the phase is computed
for all bodies but the apparent magnitudes are only computed for the planets.
The apparent magnitude is stored in the private member Magnitude of class skyObject.

In class DetailDialog I filled in the magnitude item which up to now showed
"--" for planets. Therefore I think that this addition does not break the
partial feature freeze stage we are in KDE. I have also added by hand the apparent
visual magnitude for the Sun to "-26.8".

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=378416